Create PandorasGiftBox.cs
Author Norman Lancaster - https://www.servuo.com/archive/pandoras-gift-box.582/

Usage
Use the [add command to create a PandorasGiftBox anywhere in the world. This is a container that is only accessible by Game Masters and above. It is not moveable, but is visible. I recommend creating them somewhere only staff can access.

Each gift box represents one gift that players may receive on login. Everything contained within the gift box will be duplicated and placed in the player's inventory or, if space is not available, their bank box. Gifting is subject to several conditions:

1. If the gift box is disabled (the default state) or empty, nothing will occur.
2. If the gift box is set to one gift per account (the default) a gift will be given once to the first character that logs into that account only.
3. If the gift box is not set to one gift per account, a gift will be given once to every character that logs in.

Configuration
Place items that are intended to be given as gifts into the gift box. These will be placed directly into players' packs. If it is desired to have the gifts within a container the container should be first created, the gifts placed inside, and the container placed inside the gift box. Try [add RandomGiftBox (included with this package) for some fun.

If it is desirable to give the gift to every character rather than once per account, use the [props command on the gift box and set the OnePerAccount property to false.

To enable giving of gifts from the box, use the "Activate this item" context menu item. To disable giving of gifts, use the "Deactivate this item" context menu. In this way limited-time gifts can be given without the need to permanently delete the original gift set.

Notes
Each gift box retains the data of which mobiles and accounts have received a gift from the box. If the box is deleted, the records of which mobiles have received the gift is lost. Creating a new gift box with the same contents will give a duplicate gift to those who already received one.

The nature of data retention can also be exploited. For instance, to implement a system whereby the set of gifts given to new accounts is changed every three months, simply remove the contents of the "new account" box and replace it with something new. That way pre-existing accounts will not receive the gift, but new accounts will.