Updated MonsterNests
Update includes ability to damage the nest with spells.

Important Note: The nest can only be targeted by holding Ctrl+Shift to see and target its name or drag out its healthbar and target that. The nest graphic itself can not be used to target the nest with spells.

Added Undead Nests.